refactor: split ApiModels into individual files and reformat CliConfig @SerialName
Extract all 12 model classes from ApiModels.kt into independent files
for better discoverability and maintainability. Reformat @SerialName
annotations in CliConfig.kt to two-line style for readability.
