+++ library/endpoints/src/commonMain/kotlin/com/linroid/kdown/endpoints/model/ErrorResponse.kt
@@ -1,0 +1,12 @@
+package com.linroid.kdown.endpoints.model
+
+import kotlinx.serialization.Serializable
+
+/**
+ * Generic error response body.
+ */
+@Serializable
+data class ErrorResponse(
+  val error: String,
+  val message: String,
+)
