+++ library/endpoints/src/commonMain/kotlin/com/linroid/kdown/endpoints/model/PriorityRequest.kt
@@ -1,0 +1,13 @@
+package com.linroid.kdown.endpoints.model
+
+import kotlinx.serialization.Serializable
+
+/**
+ * Request body for setting task priority.
+ *
+ * @property priority one of LOW, NORMAL, HIGH, URGENT
+ */
+@Serializable
+data class PriorityRequest(
+  val priority: String,
+)
